+++ src/Stormancer.Plugins/GameHistory/Stormancer.Server.Plugins.GameHistory/Changelog.rst
@@ -9,6 +9,13 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Add an optional parameter to the AddToHistory method for waiting for the records to be indexed
+- Rename GameHistoryRecord.players to GameHistoryRecord.Players
+
 2.0.2.1
 -------
 Changed